requirements.txt
certifi==2019.9.11
dj-database-url==0.5.0
Django==2.1
django-heroku==0.3.1
djangorestframework==3.10.3
gunicorn==19.9.0
pipenv==2018.11.26
psycopg2-binary==2.9.9
pytz==2019.3
virtualenv==16.7.7
virtualenv-clone==0.5.3
whitenoise==4.1.4
